Move pages methods tests into their own test class in the Pages folder
