+++ tests/Cms/Page/PageMethodsTest.php
@@ -2,8 +2,6 @@
 
 namespace Kirby\Cms;
 
-use Kirby\TestCase;
-
 class PageMethodsTest extends TestCase
 {
 	public function setUp(): void
@@ -11,9 +9,6 @@ class PageMethodsTest extends TestCase
 		$this->app = new App([
 			'pageMethods' => [
 				'test' => fn () => 'page method'
-			],
-			'pagesMethods' => [
-				'test' => fn () => 'pages method'
 			],
 			'site' => [
 				'children' => [
@@ -34,11 +29,5 @@ class PageMethodsTest extends TestCase
 	{
 		$page = $this->app->page('test');
 		$this->assertSame('page method', $page->test());
-	}
-
-	public function testPagesMethod()
-	{
-		$pages = $this->app->site()->children();
-		$this->assertSame('pages method', $pages->test());
 	}
 }
